changed all to postgres sql commands

+++ src/app/editor/(course)/approve/[story_id]/route.js
@@ -1,58 +1,46 @@
-import query, { update } from "lib/db";
+import { sql } from "lib/db";
 import { getToken } from "next-auth/jwt";
 import { NextResponse } from "next/server";
 
 export async function GET(req, { params: { story_id } }) {
-  try {
-    const token = await getToken({ req });
+  const token = await getToken({ req });
 
-    if (!token.role) return new Response("Error not allowed", { status: 401 });
+  if (!token.role) return new Response("Error not allowed", { status: 401 });
 
-    let answer = await set_approve({
-      story_id: parseInt(story_id),
-      user_id: token?.id,
-    });
+  let answer = await set_approve({
+    story_id: parseInt(story_id),
+    user_id: token?.id,
+  });
 
-    if (answer === undefined)
-      return new Response("Error not found", { status: 404 });
+  if (answer === undefined)
+    return new Response("Error not found", { status: 404 });
 
-    return NextResponse.json(answer);
-  } catch (err) {
-    return new Response(err.message, { status: 500 });
-  }
+  return NextResponse.json(answer);
 }
 
 async function set_status(data) {
-  return await update("story", data, ["status"]);
+  return sql`
+  UPDATE story SET ${sql(data, "status")}
+  WHERE id = ${data.id}
+`;
 }
 
 async function set_approve({ story_id, user_id }) {
-  let res = await query(
-    `SELECT id FROM story_approval WHERE story_id = ? AND user_id = ?;`,
-    [story_id, user_id],
-  );
+  let res =
+    await sql`SELECT id FROM story_approval WHERE story_id = ${story_id} AND user_id = ${user_id};`;
   let action;
   if (res.length) {
-    await query(
-      `DELETE FROM story_approval WHERE story_id = ? AND user_id = ?;`,
-      [story_id, user_id],
-    );
+    await sql`DELETE FROM story_approval WHERE story_id = ${story_id} AND user_id = ${user_id};`;
     action = "deleted";
   } else {
-    await query(
-      `INSERT INTO story_approval (story_id, user_id) VALUES (?, ?);`,
-      [story_id, user_id],
-    );
+    await sql`INSERT INTO story_approval (story_id, user_id) VALUES (${story_id}, ${user_id});`;
     action = "added";
   }
   let res2 = (
-    await query(
-      `SELECT COUNT(id) as count FROM story_approval WHERE story_id = ?;`,
-      [story_id],
-    )
+    await sql`SELECT COUNT(id) as count FROM story_approval WHERE story_id = ${story_id};`
   )[0];
-  let count = res2["count"];
-
+  let count = parseInt(res2.count);
+  console.log(count);
   let status = undefined;
   if (count === 0) status = "draft";
   if (count === 1) status = "feedback";
@@ -61,21 +49,16 @@ async function set_approve({ story_id, user_id }) {
   await set_status({ status: status, id: story_id });
 
   // get the number of finished stories in this set
-  let res3 = await query(
-    `SELECT story.id, story.public FROM story WHERE set_id = (SELECT set_id FROM story WHERE id = ? LIMIT 1) AND
-                                            course_id = (SELECT course_id FROM story WHERE id = ? LIMIT 1) AND status = 'finished' AND deleted = 0;`,
-    [story_id, story_id],
-  );
+  let res3 =
+    await sql`SELECT story.id, story.public FROM story WHERE set_id = (SELECT set_id FROM story WHERE id = ${story_id} LIMIT 1) AND
+                                            course_id = (SELECT course_id FROM story WHERE id = ${story_id} LIMIT 1) AND status = 'finished' AND NOT deleted;`;
 
   let published = [];
   if (res3.length >= 4) {
     let date_published = new Date().toISOString();
     for (let story of res3) {
       if (story.public === 0) {
-        await query(
-          `UPDATE story SET public = 1, date_published = ? WHERE id = ?;`,
-          [date_published, story.id],
-        );
+        await sql`UPDATE story SET public = true, date_published = ${date_published} WHERE id = ${story.id};`;
         published.push(story.id);
       }
     }
